+++ killer_instagram/src/routes/users.py
@@ -17,7 +17,7 @@ router = APIRouter(prefix='/users', tags=['users'])
 security = HTTPBearer()
 
 allowd_operation = RoleRights(["user", "moderator", "admin"])
-allowd_operation_ban = RoleRights(["admin"])
+allowd_operation_by_admin = RoleRights(["admin"])
 
 
 @router.get('/',
@@ -95,7 +95,7 @@ async def get_user_profile(username,
                     "banned": user.banned,
                     "user role":user.role,
                     "avatar URL": user.avatar,
-                    "quantity_of_loaded_images": quantity_of_loaded_images
+                    "quantity_of_loaded_images": quantity_of_loaded_images,
                     }
 
     return user_profile
@@ -135,7 +135,7 @@ async def update_avatar_user(file: UploadFile = File(),
                         response_model=UserResponce, 
                         status_code=status.HTTP_200_OK,
                         dependencies=[Depends(logout_dependency), 
-                                      Depends(allowd_operation_ban)],
+                                      Depends(allowd_operation_by_admin)],
                         )
 async def update_banned_status(user_id:str,
                                 credentials: HTTPAuthorizationCredentials = Security(security), 
@@ -169,7 +169,7 @@ async def update_banned_status(user_id:str,
                         response_model=UserResponce, 
                         status_code=status.HTTP_200_OK,
                         dependencies=[Depends(logout_dependency), 
-                                      Depends(allowd_operation_ban)],
+                                      Depends(allowd_operation_by_admin)],
                         )
 async def update_unbanned_status(user_id:str,
                                 credentials: HTTPAuthorizationCredentials = Security(security), 
@@ -198,3 +198,34 @@ async def update_unbanned_status(user_id:str,
 
     await repository_users.update_unbanned_status(user, db)
     return user
+
+@router.delete('/{user_id}',
+               status_code=status.HTTP_200_OK,
+               dependencies=[Depends(logout_dependency), 
+                             Depends(allowd_operation_by_admin)])
+async def delete_user(user_id: int, current_user: User = Depends(service_auth.get_current_user),
+                      db: Session = Depends(get_db)):
+    """
+    The delete_user function allows an admin to delete a user.
+
+    :param user_id: int: ID of the user to be deleted
+    :param current_user: User: Get the current user from the database
+    :param db: Session: Database session
+    :return: HTTP status 204 (No Content)
+    """
+ 
+    user = await repository_users.get_user_by_id(user_id, db)
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+      
+    if current_user.id != 1 and user.role == "admin":
+        raise HTTPException(status_code=403, detail="Permission denied. Only supreadmin can delede other admin.")
+    
+    if current_user.id == user_id:
+        raise HTTPException(status_code=403, detail="Permission denied. Cannot delete own account.")
+    
+    if user.id == 1:
+        raise HTTPException(status_code=403, detail="Permission denied. Superadmin user cannot be deleted.")
+
+    await repository_users.delete_user(user_id, db)
+    return {"message": f"User {current_user.email} successfully deleted"}
